+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Batch rendering of pieces in the receive form. Refs UIREC-461.
 * ECS - Fetch locations, holdings, and items from all related tenants. Refs UIREC-467.
 * Hide number generator icon in Receiving when number generators are off. Refs UIREC-460.
+* Improve display in receiving. Refs UIREC-464.
 
 ## [7.0.2](https://github.com/folio-org/ui-receiving/tree/v7.0.2) (2025-04-14)
 [Full Changelog](https://github.com/folio-org/ui-receiving/compare/v7.0.1...v7.0.2)